Reorganize unit tests.
